Rename to _base so it doesn't get loaded as a template file.
